# Optimize with getPath; always try to use paths to get and update