--- a/packages/slate/src/models/node.js
+++ b/packages/slate/src/models/node.js
@@ -2140,7 +2140,6 @@ memoize(Node.prototype, [
   'getNode',
   'getNodeAtPath',
   'getOffset',
-  'getOffsetAtRange',
   'getParent',
   'getPath',
   'getPathAsString',