--- a/packages/slate/src/models/node.js
+++ b/packages/slate/src/models/node.js
@@ -1761,7 +1761,7 @@ class Node {
    */
 
   hasNode(key) {
-    return Array.isArray(this.getPath(key))
+    return !!this.getPath(key)
   }
 
   /**