This is a part of #1771 .  This PR provides a getPathAsString to take advantage from the speed of string concat, and also try to use path to get, search and update