@ianstormtaylor 2018-05-02
@zhujinxuan this change looks like too much magic for the sake of performance to me that will actively make the codebase harder to understand. Encoding paths as strings, parsing them as integers, etc. It all seems like a convoluted way to get this.

If I understand correctly, all of this would be unnecessary if we used paths in the selections instead of (or in addition to) keys?

I'd rather not take on a lot of complicated logic if it could be avoided by making a better architectural change to the library.