@zhujinxuan 2018-05-02
@ianstormtaylor We still have a great (not such great) improvement if we use getPath() in getAncestors, getFurthestAncestor, getFurthest...  I use string just for an extra 50% improvement, while only use getPath without string magic can have 800% already. 

I will remove the string magic and submit commits.