@zhujinxuan 2018-05-02
@ianstormtaylor Indeed, it is related to selection/range.  It is alike: 
1. Do not do DFS twice
2. If it is possible, do not create an Immutable Object. Because Immutable takes a longer time to get itself created.

All are switched to getPath to ensure DFS is not duplicated.  And creating array is faster than creating Immutable List.