@benjycui 2018-05-03
Hope that you can hold this PR.

Actually, I had done something like this in our forked slate, but I cannot PR now for I have to apply first.

After my PR, the benchmark(base on https://github.com/ianstormtaylor/slate/pull/1765):

```bash
benchmarks
    changes
      delete-backward
        user : 530.27 → 819.12 ops/sec (54.4720511289997% faster)
        hr : 912.60 → 1287.85 ops/sec (41.11925192135757% faster)
      delete-forward
        user : 142.23 → 228.58 ops/sec (60.7065909325922% faster)
        hr : 235.44 → 366.03 ops/sec (55.46557149243505% faster)
      insert-node-by-key
        user : 314.85 → 807.03 ops/sec (156.31846797327427% faster) 🙌
        hr : 538.00 → 1243.08 ops/sec (131.05496517171622% faster) 🙌
      insert-text-by-key-multiple
        user : 79.64 → 190.47 ops/sec (139.1550052283623% faster) 🙌
        hr : 114.91 → 301.94 ops/sec (162.75320053456548% faster) 🙌
      insert-text-by-key
        user : 576.61 → 1672.88 ops/sec (190.1241191631047% faster) 🙌
        hr : 896.01 → 2816.23 ops/sec (214.30660136689013% faster) 🙌
      insert-text
        user : 2377.44 → 2940.70 ops/sec
        hr : 4158.32 → 5221.40 ops/sec
      normalize
        user : 882.67 → 931.35 ops/sec
        hr : 1433.59 → 1418.38 ops/sec
      remove-node-by-key
        user : 415.29 → 970.73 ops/sec (133.74660793652228% faster) 🙌
        hr : 685.31 → 1552.40 ops/sec (126.5232320035702% faster) 🙌
      split-block
        user : 556.96 → 691.33 ops/sec
        hr : 886.48 → 1043.11 ops/sec
    models
      from-json-big
        user : 4.70 → 4.75 ops/sec
        hr : 6.66 → 6.74 ops/sec
      from-json
        user : 209.83 → 186.56 ops/sec
        hr : 380.03 → 340.77 ops/sec
      get-active-marks-at-range
        user : 480.36 → 449.10 ops/sec
        hr : 639.31 → 544.03 ops/sec
      get-ancestors
        user : 5472.29 → 30735.45 ops/sec (461.65649037140025% faster) 🙌
        hr : 8218.20 → 74928.36 ops/sec (811.7364539092894% faster) 🙌
      get-blocks-at-range
        user : 905.50 → 1568.01 ops/sec (73.16469292465477% faster)
        hr : 1796.80 → 2832.79 ops/sec (57.65788195885708% faster)
      get-blocks
        user : 8456.70 → 8796.07 ops/sec
        hr : 13770.19 → 12915.79 ops/sec
      get-characters-at-range
        user : 304.38 → 283.00 ops/sec
        hr : 387.57 → 333.03 ops/sec
      get-characters
        user : 4276.38 → 3765.22 ops/sec
        hr : 5671.35 → 5117.25 ops/sec
      get-common-ancestor
        user : 1172.95 → 31975.12 ops/sec (2626.039633262033% faster) 😱
        hr : 1527.20 → 60860.45 ops/sec (3885.089857942521% faster) 😱
      get-furthest-ancestor
        user : 9174.95 → 98356.25 ops/sec (972.0089992082658% faster) 🙌
        hr : 14504.19 → 259868.42 ops/sec (1691.67872889726% faster) 😱
      get-inlines-at-range
        user : 734.96 → 2373.07 ops/sec (222.88369262836073% faster) 🙌
        hr : 925.96 → 3385.41 ops/sec (265.61168992987535% faster) 🙌
      get-inlines
        user : 11932.45 → 11690.45 ops/sec
        hr : 19018.06 → 18908.81 ops/sec
      get-leaves
        user : 10595.06 → 10224.20 ops/sec
        hr : 20530.07 → 18697.95 ops/sec
      get-marks-at-range
        user : 199.28 → 223.86 ops/sec
        hr : 287.74 → 300.91 ops/sec
      get-marks
        user : 735.03 → 710.85 ops/sec
        hr : 1082.48 → 1111.09 ops/sec
      get-parent
        user : 6711.79 → 34013.90 ops/sec (406.7785689207298% faster) 🙌
        hr : 8569.09 → 80001.76 ops/sec (833.6088349983935% faster) 🙌
      get-path
        user : 2400.40 → 256104.97 ops/sec (10569.271371374982% faster) 😱
        hr : 3499.55 → 579564.64 ops/sec (16461.107761650317% faster) 😱
      get-texts-at-range
        user : 4371.11 → 6576.90 ops/sec (50.46289262370672% faster)
        hr : 6499.21 → 10157.09 ops/sec (56.28180427844705% faster)
      get-texts
        user : 10660.73 → 11297.84 ops/sec
        hr : 15847.01 → 18445.46 ops/sec
      getDescendant
        user : 9553.02 → 44818.60 ops/sec (369.1565709271796% faster) 🙌
        hr : 14032.07 → 131351.23 ops/sec (836.0788587363576% faster) 🙌
      has-node-multiple
        user : 2093.88 → 9774.22 ops/sec (366.7990937667193% faster) 🙌
        hr : 2386.29 → 12934.66 ops/sec (442.0398514115642% faster) 🙌
      has-node
        user : 9871.90 → 44356.45 ops/sec (349.3205556074771% faster) 🙌
        hr : 14706.70 → 119986.85 ops/sec (715.8654576617625% faster) 🙌
      to-json
        user : 1013.53 → 627.81 ops/sec (61% slower)
        hr : 1465.05 → 898.23 ops/sec (63% slower)
      update-node
        user : 3724.43 → 6708.29 ops/sec (80.11586139756857% faster)
        hr : 5524.59 → 11372.22 ops/sec (105.8473597097442% faster) 🙌
    html-serializer
      deserialize
        user : 9.81 → 25.96 ops/sec (164.7090172892424% faster) 🙌
        hr : 13.10 → 33.85 ops/sec (158.2912669871731% faster) 🙌
      serialize
        user : 298.81 → 316.10 ops/sec
        hr : 493.48 → 488.24 ops/sec
    plain-serializer
      deserialize
        user : 489.34 → 501.56 ops/sec
        hr : 814.25 → 836.74 ops/sec
      serialize
        user : 2587.39 → 2458.94 ops/sec
        hr : 3938.02 → 3678.04 ops/sec
    rendering
      normal
        user : 88.65 → 76.63 ops/sec
        hr : 120.21 → 99.16 ops/sec
```