@benjycui 2018-05-03
My solution will not assume key is number string or sequential, and it works fine without memoization(which may cause memory leaks https://github.com/ianstormtaylor/slate/issues/1402)