@zhujinxuan 2018-05-03
@benjycui I have just added you to my repo, so you can edit on branches directly.  Can you add your `get-ancestors`, `get-parent` and other benches to the branch of #1765 (benchmark-consistence branch)?

Thank you.