@zhujinxuan 2018-05-03
@benjycui I think it is fine to hold this PR.  
Can you comment on the code panel if you feel something is not ideal?  

I am very surprised about your getPath optimization (100x faster).  Can I know what your solution is?