@benjycui 2018-05-04
Ref: https://github.com/ianstormtaylor/slate/pull/1818