@zhujinxuan 2018-07-05
Hi, there are already too much changes in node.js and operation.js.  I think it is not optimistic to wait for @benjycui  #1818 .  I will submit this PR this weekend with resolving the node conflict.  

For the path refactor, I will do it in future PR with a different interface.