@zhujinxuan 2018-07-05
@ianstormtaylor The conflict is resolved.  Can you have a review on this branch?

Thank you,
Jinxuan Zhu